Fin de l'implémentation des différentes fonctions

+++ projet/communitymanager/views.py
@@ -2,17 +2,26 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from .models import Communaute
 from .models import Post
+from .models import Commentaire
 from .forms import NewPostForm
+from .forms import NewCommentForm
 from datetime import datetime
 
 
 def log(request):
-    return redirect('newsfeed')
+    return redirect('communautes')
 
 
 @login_required()
 def newsfeed(request):
-    return render(request, 'communitymanager/newsfeed.html')
+    communautes = request.user.communaute_set.all()
+    list = []
+    for com in communautes:
+        l = Post.objects.filter(communaute=com)
+        for p in l:
+            list.append(p)
+
+    return render(request, 'communitymanager/newsfeed.html', locals())
 
 
 def disconnect(request):
@@ -41,30 +50,31 @@ def abonnement(request, com_id):
 
 @login_required()
 def communaute(request, com_id):
-    list = Post.objects.all()
+    list = Post.objects.filter(communaute=Communaute.objects.get(id=com_id))
     com = Communaute.objects.get(id=com_id)
     return render(request, 'communitymanager/communaute.html', locals())
 
 
 @login_required()
-def post_details(request, post_id):
+def post(request, post_id):
     post = Post.objects.get(id=post_id)
+    comments = Commentaire.objects.filter(post = post)
+    form = NewCommentForm(request.POST or None)
+    if form.is_valid():
+        commentaire = form.save(commit=False)
+        commentaire.date_creation = datetime.now()
+        commentaire.auteur = request.user
+        commentaire.post = post
+        commentaire.save()
+        return redirect('post', post_id)
     return render(request, 'communitymanager/post-details.html', locals())
 
 
 @login_required()
-def create_post(request):
+def nouveau_post(request):
     form = NewPostForm(request.POST or None)
     if form.is_valid():
-        post = Post()
-        print(post)
-        post.titre = form.cleaned_data['titre']
-        post.description = form.cleaned_data['description']
-        post.communaute = form.cleaned_data['communaute']
-        post.priorite = form.cleaned_data['priorite']
-        post.evenementiel = form.cleaned_data['evenementiel']
-        if post.evenementiel:
-            post.date_evenement = form.cleaned_data['date_evenementiel']
+        post = form.save(commit=False)
         post.date_creation = datetime.now()
         post.auteur = request.user
         post.save()
@@ -76,21 +86,31 @@ def create_post(request):
 
 @login_required()
 def modif_post(request, post_id):
-    form = NewPostForm(request.POST or None)
-    if form.is_valid():
-        post = form.save(commit=False)
-        post.auteur = request.user
-        post.date_creation = datetime.now()
-        return redirect('details', post_id=post_id)
-    else:
-        p = get_object_or_404(Post, id=post_id)
-        form = NewPostForm(instance=p)
-        return render(request, 'communitymanager/modify-post.html', locals())
+    p = get_object_or_404(Post, id=post_id)
+    if p.auteur == request.user:
+        form = NewPostForm(request.POST or None)
+        if form.is_valid():
+            post = form.save(commit=False)
+            post.auteur = request.user
+            post.date_creation = datetime.now()
+            return redirect('post', post_id=post_id)
+        else:
+            form = NewPostForm(instance=p)
+            return render(request, 'communitymanager/modify-post.html', locals())
+    else :
+        list = Post.objects.filter(communaute=Communaute.objects.get(id=Post.objects.get(id=post_id).communaute.id))
+        can_modif = True
+        return render(request, 'communitymanager/communaute.html', locals())
 
 
 @login_required()
 def suppress_post(request, post_id):
-    post = Post.objects.get(id=post_id)
-    com_id = post.communaute.id
-    post.delete()
-    return redirect('communaute', com_id=com_id)
+    post = get_object_or_404(Post, id=post_id)
+    if post.auteur == request.user:
+        com_id = post.communaute.id
+        post.delete()
+        return redirect('communaute', com_id=com_id)
+    else :
+        list = Post.objects.filter(communaute=Communaute.objects.get(id=Post.objects.get(id=post_id).communaute.id))
+        can_suppress = True
+        return render(request, 'communitymanager/communaute.html', locals())
